Change tickstillupdate for Copper Pipes/Wires to 1. That means that they are updated each tick.I did this because most other mods have a max value they can accept/send per tick.Ex Redstone Energy Cell can accept 2000 RF/t and our wire can do 2500 RF/t but it did it only two ticks, so for the Energy Cell it was actually trying to insert 5000 RF every 2 ticks and it could only insert 2000 RF to ticks, so that means that the max rate was only 1000 RF/t, altough the wire is more efficient in terms of performance.
The same happens with the pipe, cause it updated every 4 ticks, it actually tried
inserting 400 mB on one tick, and with the BC Golden pipes it only did 40 mb
per 4 ticks, which is 10 mb/t, much less that the actual pipes.

I still left the value there, maybe we can make it a config onption for users
who want more performance but less speed.